+++ apps/router-e2e/__e2e__/03-server-actions-only/components/server-actions.tsx
@@ -1,16 +1,19 @@
 'use server';
 
 import Ionicons from '@expo/vector-icons/Ionicons';
-import { Image, Text } from 'react-native';
 import { Link, Stack } from 'expo-router';
+import { unstable_headers } from 'expo-router/rsc/headers';
+import { Image, Text } from 'react-native';
 
 export async function renderPage(params: { title: string }) {
+  const platformHeader = (await unstable_headers()).get('expo-platform');
   return (
     <>
       <Stack.Screen options={{ title: 'Server Actions' }} />
       <Text testID="index-text">Platform: {process.env.EXPO_OS}</Text>
       <Text testID="secret-text">Secret: {process.env.TEST_SECRET_VALUE}</Text>
       <Text>Render: {Date.now()}</Text>
+      <Text testID="platform-header">Header: {platformHeader}</Text>
 
       <Image
         testID="main-image"